feat: Add shuffle button for new game
Introduces a shuffle button to restart the current game, providing a quick way to re-deal cards and begin a new round. This option is only available when the user is not acting as a client in a multiplayer session. Also, a `Shuffle` icon from `lucide-react` was added to the available icons.